day_4/task_3/requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
python-dotenv==1.0.0
lxml==4.9.3 